app/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Pillow==9.0.0
